refactor: extract document chunking from CitationService (Closes aliwatters/gsuite-mcp#103)
Move all chunking logic, CreateIndex, and RefreshIndex out of the 716-line
citation_service.go into dedicated files:
- chunker.go: DocumentChunker interface, chunkPptxData, extractTextFromXML,
  chunkID, chunkText, extractSlidesText, downloadAndChunk, limitedRead
- index_create.go: CreateIndex method
- index_refresh.go: RefreshIndex, removeFileFromIndex, reindexFile

citation_service.go reduced from 716 to 356 lines. All tests pass.
