+++ internal/citation/citation_service.go
@@ -1,19 +1,11 @@
 package citation
 
 import (
-	"archive/zip"
-	"bytes"
 	"context"
-	"crypto/sha256"
-	"encoding/xml"
 	"fmt"
-	"io"
 	"net/http"
-	"sort"
-	"strconv"
 	"strings"
 	"sync"
-	"time"
 
 	"google.golang.org/api/drive/v3"
 	"google.golang.org/api/option"
@@ -77,9 +69,6 @@ type IndexEntry struct {
 	SheetID string `json:"sheet_id"`
 }
 
-// maxExportSize limits Drive export downloads to 50MB.
-const maxExportSize = 50 * 1024 * 1024
-
 // NewRealCitationService creates a new citation service from an HTTP client.
 func NewRealCitationService(ctx context.Context, client *http.Client, cfg *CitationConfig) (*RealCitationService, error) {
 	driveSrv, err := drive.NewService(ctx, option.WithHTTPClient(client))
@@ -138,77 +127,6 @@ func (s *RealCitationService) getStore(ctx context.Context, indexID string) (*Du
 	return store, nil
 }
 
-func (s *RealCitationService) CreateIndex(ctx context.Context, name, folderID string) (*IndexInfo, error) {
-	// Create spreadsheet with required tabs
-	ss := &sheets.Spreadsheet{
-		Properties: &sheets.SpreadsheetProperties{Title: "Citation Index: " + name},
-		Sheets: []*sheets.Sheet{
-			{Properties: &sheets.SheetProperties{Title: "chunks"}},
-			{Properties: &sheets.SheetProperties{Title: "concepts"}},
-			{Properties: &sheets.SheetProperties{Title: "summaries"}},
-			{Properties: &sheets.SheetProperties{Title: "files"}},
-			{Properties: &sheets.SheetProperties{Title: "metadata"}},
-		},
-	}
-	created, err := s.sheetsService.Spreadsheets.Create(ss).Context(ctx).Do()
-	if err != nil {
-		return nil, fmt.Errorf("creating index sheet: %w", err)
-	}
-
-	// Move to folder if specified
-	if folderID != "" {
-		_, err = s.driveService.Files.Update(created.SpreadsheetId, nil).
-			AddParents(folderID).
-			SupportsAllDrives(true).
-			Context(ctx).Do()
-		if err != nil {
-			return nil, fmt.Errorf("moving sheet to folder: %w", err)
-		}
-	}
-
-	// Write headers
-	headers := map[string][][]any{
-		"chunks!A1":    {{"id", "file_id", "file_name", "content", "summary", "concepts", "page_number", "section_heading", "paragraph_index", "char_start", "char_end"}},
-		"concepts!A1":  {{"concept", "chunk_ids"}},
-		"summaries!A1": {{"level", "parent_id", "summary"}},
-		"files!A1":     {{"file_id", "file_name", "mime_type", "modified_time", "chunk_count"}},
-		"metadata!A1":  {{"key", "value"}},
-	}
-	var data []*sheets.ValueRange
-	for r, v := range headers {
-		data = append(data, &sheets.ValueRange{Range: r, Values: v})
-	}
-	_, err = s.sheetsService.Spreadsheets.Values.BatchUpdate(created.SpreadsheetId, &sheets.BatchUpdateValuesRequest{
-		ValueInputOption: "RAW",
-		Data:             data,
-	}).Context(ctx).Do()
-	if err != nil {
-		return nil, fmt.Errorf("writing headers: %w", err)
-	}
-
-	// Write initial metadata
-	now := time.Now().UTC().Format(time.RFC3339)
-	metaRows := [][]any{
-		{"index_id", name},
-		{"source_folder_id", folderID},
-		{"created_at", now},
-		{"doc_count", "0"},
-		{"chunk_count", "0"},
-	}
-	_, err = s.sheetsService.Spreadsheets.Values.Append(created.SpreadsheetId, "metadata!A2", &sheets.ValueRange{
-		Values: metaRows,
-	}).ValueInputOption("RAW").Context(ctx).Do()
-	if err != nil {
-		return nil, fmt.Errorf("writing metadata: %w", err)
-	}
-
-	return &IndexInfo{
-		IndexID:        name,
-		SheetID:        created.SpreadsheetId,
-		SourceFolderID: folderID,
-		CreatedAt:      now,
-	}, nil
-}
 
 func (s *RealCitationService) AddDocuments(ctx context.Context, indexID string, fileIDs []string) (int, error) {
 	store, err := s.getStore(ctx, indexID)
@@ -249,6 +167,7 @@ func (s *RealCitationService) AddDocuments(ctx context.Context, indexID string, 
 	return totalChunks, nil
 }
 
+// chunkFile dispatches to the appropriate chunking strategy based on MIME type.
 func (s *RealCitationService) chunkFile(ctx context.Context, file *drive.File) ([]Chunk, error) {
 	switch file.MimeType {
 	case "application/vnd.google-apps.presentation":
@@ -258,7 +177,7 @@ func (s *RealCitationService) chunkFile(ctx context.Context, file *drive.File) (
 		return s.chunkDownloadedFile(ctx, file)
 	default:
 		if strings.HasPrefix(file.MimeType, "application/vnd.google-apps.") {
-			// Google-native file — export as text
+			// Google-native file — export as plain text
 			return s.chunkExportedText(ctx, file)
 		}
 		// Non-Google file — try downloading raw content
@@ -266,57 +185,17 @@ func (s *RealCitationService) chunkFile(ctx context.Context, file *drive.File) (
 	}
 }
 
+// chunkSlides extracts per-slide text from a Google Slides presentation via the Slides API.
 func (s *RealCitationService) chunkSlides(ctx context.Context, file *drive.File) ([]Chunk, error) {
 	pres, err := s.slidesService.Presentations.Get(file.Id).Context(ctx).Do()
 	if err != nil {
 		return nil, err
 	}
-
-	var chunks []Chunk
-	for i, slide := range pres.Slides {
-		var textParts []string
-		for _, elem := range slide.PageElements {
-			if elem.Shape != nil && elem.Shape.Text != nil {
-				for _, te := range elem.Shape.Text.TextElements {
-					if te.TextRun != nil {
-						textParts = append(textParts, te.TextRun.Content)
-					}
-				}
-			}
-			if elem.Table != nil {
-				for _, row := range elem.Table.TableRows {
-					for _, cell := range row.TableCells {
-						if cell.Text != nil {
-							for _, te := range cell.Text.TextElements {
-								if te.TextRun != nil {
-									textParts = append(textParts, te.TextRun.Content)
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-
-		content := strings.TrimSpace(strings.Join(textParts, ""))
-		if content == "" {
-			continue
-		}
-
-		chunks = append(chunks, Chunk{
-			ID:       chunkID(file.Id, i),
-			FileID:   file.Id,
-			FileName: file.Name,
-			Content:  content,
-			Location: Location{PageNumber: i + 1},
-		})
-	}
-	return chunks, nil
+	return extractSlidesText(file.Id, file.Name, pres), nil
 }
 
 // chunkDownloadedFile downloads a non-Google-native file and extracts text.
-// For .pptx files, it reads text from the XML entries inside the zip.
-// For other binary formats, it returns a single chunk with available text.
+// For .pptx files, reads text from XML inside the zip. For others, treats content as plain text.
 func (s *RealCitationService) chunkDownloadedFile(ctx context.Context, file *drive.File) ([]Chunk, error) {
 	resp, err := s.driveService.Files.Get(file.Id).SupportsAllDrives(true).Context(ctx).Download()
 	if err != nil {
@@ -324,112 +203,21 @@ func (s *RealCitationService) chunkDownloadedFile(ctx context.Context, file *dri
 	}
 	defer resp.Body.Close()
 
-	data, err := io.ReadAll(io.LimitReader(resp.Body, maxExportSize))
+	data, err := limitedRead(resp.Body)
 	if err != nil {
 		return nil, fmt.Errorf("reading %s: %w", file.Name, err)
 	}
 
-	if file.MimeType == "application/vnd.openxmlformats-officedocument.presentationml.presentation" {
-		return s.chunkPptxBytes(file, data)
-	}
-
-	// Fallback: treat as plain text
-	text := strings.TrimSpace(string(data))
-	if text == "" {
-		return nil, nil
-	}
-	return chunkText(file.Id, file.Name, text), nil
+	return downloadAndChunk(data, file)
 }
 
-// chunkPptxBytes extracts text from a pptx file (zip containing XML slides).
-// Each slide/N.xml becomes a separate chunk.
+// chunkPptxBytes extracts text from a .pptx file (zip containing XML slides).
+// Delegates to the package-level chunkPptxData function.
 func (s *RealCitationService) chunkPptxBytes(file *drive.File, data []byte) ([]Chunk, error) {
-	r, err := zip.NewReader(bytes.NewReader(data), int64(len(data)))
-	if err != nil {
-		return nil, fmt.Errorf("reading pptx zip: %w", err)
-	}
-
-	// Collect slide XML files (ppt/slides/slide1.xml, slide2.xml, ...)
-	type slideEntry struct {
-		num  int
-		file *zip.File
-	}
-	var slides []slideEntry
-	for _, f := range r.File {
-		if !strings.HasPrefix(f.Name, "ppt/slides/slide") || !strings.HasSuffix(f.Name, ".xml") {
-			continue
-		}
-		// Extract slide number from "ppt/slides/slide12.xml"
-		numStr := strings.TrimPrefix(f.Name, "ppt/slides/slide")
-		numStr = strings.TrimSuffix(numStr, ".xml")
-		num, err := strconv.Atoi(numStr)
-		if err != nil {
-			continue
-		}
-		slides = append(slides, slideEntry{num: num, file: f})
-	}
-
-	// Sort by slide number
-	sort.Slice(slides, func(i, j int) bool { return slides[i].num < slides[j].num })
-
-	var chunks []Chunk
-	for _, s := range slides {
-		text, err := extractTextFromXML(s.file)
-		if err != nil {
-			continue
-		}
-		text = strings.TrimSpace(text)
-		if text == "" {
-			continue
-		}
-		chunks = append(chunks, Chunk{
-			ID:       chunkID(file.Id, s.num-1),
-			FileID:   file.Id,
-			FileName: file.Name,
-			Content:  text,
-			Location: Location{PageNumber: s.num},
-		})
-	}
-	return chunks, nil
+	return chunkPptxData(file, data)
 }
 
-// extractTextFromXML reads a zip entry and extracts all text content from XML <a:t> tags.
-func extractTextFromXML(f *zip.File) (string, error) {
-	rc, err := f.Open()
-	if err != nil {
-		return "", err
-	}
-	defer rc.Close()
-
-	decoder := xml.NewDecoder(rc)
-	var textParts []string
-	var inText bool
-	for {
-		tok, err := decoder.Token()
-		if err != nil {
-			break
-		}
-		switch t := tok.(type) {
-		case xml.StartElement:
-			// <a:t> contains text in OOXML presentations
-			if t.Name.Local == "t" {
-				inText = true
-			}
-		case xml.EndElement:
-			if t.Name.Local == "t" {
-				inText = false
-			}
-		case xml.CharData:
-			if inText {
-				textParts = append(textParts, string(t))
-			}
-		}
-	}
-	return strings.Join(textParts, " "), nil
-}
-
-// chunkExportedText exports a file as plain text via Drive API and chunks by paragraphs.
-// Handles Google Docs, Sheets, and any other exportable format.
+// chunkExportedText exports a Google-native file as plain text and splits into chunks.
 func (s *RealCitationService) chunkExportedText(ctx context.Context, file *drive.File) ([]Chunk, error) {
 	resp, err := s.driveService.Files.Export(file.Id, "text/plain").Context(ctx).Download()
 	if err != nil {
@@ -437,7 +225,7 @@ func (s *RealCitationService) chunkExportedText(ctx context.Context, file *drive
 	}
 	defer resp.Body.Close()
 
-	data, err := io.ReadAll(io.LimitReader(resp.Body, maxExportSize))
+	data, err := limitedRead(resp.Body)
 	if err != nil {
 		return nil, err
 	}
@@ -565,152 +353,4 @@ func (s *RealCitationService) FormatCitation(_ context.Context, chunk Chunk) str
 	return "Source: " + strings.Join(parts, ", ")
 }
 
-func (s *RealCitationService) RefreshIndex(ctx context.Context, indexID string) (*RefreshResult, error) {
-	store, err := s.getStore(ctx, indexID)
-	if err != nil {
-		return nil, err
-	}
 
-	// Get currently indexed files
-	indexed, err := store.GetIndexedFiles(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("getting indexed files: %w", err)
-	}
-
-	result := &RefreshResult{}
-	indexedMap := make(map[string]IndexedFile, len(indexed))
-	for _, f := range indexed {
-		indexedMap[f.FileID] = f
-	}
-
-	// Check each indexed file against Drive
-	for _, prev := range indexed {
-		current, err := s.driveService.Files.Get(prev.FileID).Fields("id,name,mimeType,modifiedTime,trashed").SupportsAllDrives(true).Context(ctx).Do()
-		if err != nil {
-			result.Errors = append(result.Errors, fmt.Sprintf("%s: %v", prev.FileName, err))
-			continue
-		}
-
-		// File was trashed → remove
-		if current.Trashed {
-			if delErr := s.removeFileFromIndex(ctx, store, prev.FileID); delErr != nil {
-				result.Errors = append(result.Errors, fmt.Sprintf("removing %s: %v", prev.FileName, delErr))
-			} else {
-				result.Removed = append(result.Removed, prev.FileName)
-			}
-			continue
-		}
-
-		// File was renamed
-		if current.Name != prev.FileName {
-			result.Renamed = append(result.Renamed, fmt.Sprintf("%s → %s", prev.FileName, current.Name))
-		}
-
-		// File was modified → re-chunk
-		if current.ModifiedTime != prev.ModifiedTime {
-			if reErr := s.reindexFile(ctx, store, current); reErr != nil {
-				result.Errors = append(result.Errors, fmt.Sprintf("re-indexing %s: %v", current.Name, reErr))
-			} else {
-				result.Updated = append(result.Updated, current.Name)
-			}
-		}
-	}
-
-	return result, nil
-}
-
-// removeFileFromIndex deletes a file's chunks and tracking record.
-func (s *RealCitationService) removeFileFromIndex(ctx context.Context, store *DualStore, fileID string) error {
-	if err := store.DeleteChunksByFileID(ctx, fileID); err != nil {
-		return fmt.Errorf("deleting chunks for file %s: %w", fileID, err)
-	}
-	return store.DeleteIndexedFile(ctx, fileID)
-}
-
-// reindexFile removes old chunks and re-chunks the file.
-func (s *RealCitationService) reindexFile(ctx context.Context, store *DualStore, file *drive.File) error {
-	// Remove old chunks
-	if err := store.DeleteChunksByFileID(ctx, file.Id); err != nil {
-		return fmt.Errorf("deleting old chunks: %w", err)
-	}
-
-	// Re-chunk
-	chunks, err := s.chunkFile(ctx, file)
-	if err != nil {
-		return fmt.Errorf("chunking: %w", err)
-	}
-
-	if err := store.SaveChunks(ctx, chunks); err != nil {
-		return fmt.Errorf("saving chunks: %w", err)
-	}
-
-	// Update tracking
-	return store.SaveIndexedFile(ctx, IndexedFile{
-		FileID:       file.Id,
-		FileName:     file.Name,
-		MimeType:     file.MimeType,
-		ModifiedTime: file.ModifiedTime,
-		ChunkCount:   len(chunks),
-	})
-}
-
-// chunkID generates a deterministic chunk ID from file ID and offset.
-func chunkID(fileID string, offset int) string {
-	h := sha256.Sum256([]byte(fmt.Sprintf("%s:%d", fileID, offset)))
-	return fmt.Sprintf("%x", h[:8])
-}
-
-// chunkText splits text into chunks by paragraph boundaries.
-// Targets ~500-1000 tokens (~2000-4000 chars) per chunk.
-func chunkText(fileID, fileName, text string) []Chunk {
-	const maxChunkSize = 3000
-
-	paragraphs := strings.Split(text, "\n\n")
-	var chunks []Chunk
-	var current strings.Builder
-	paraIdx := 0
-	charPos := 0
-	chunkStart := 0
-
-	flush := func() {
-		content := strings.TrimSpace(current.String())
-		if content == "" {
-			return
-		}
-		chunks = append(chunks, Chunk{
-			ID:       chunkID(fileID, len(chunks)),
-			FileID:   fileID,
-			FileName: fileName,
-			Content:  content,
-			Location: Location{
-				ParagraphIndex: paraIdx,
-				CharStart:      chunkStart,
-				CharEnd:        charPos,
-			},
-		})
-		current.Reset()
-		chunkStart = charPos
-	}
-
-	for _, para := range paragraphs {
-		para = strings.TrimSpace(para)
-		if para == "" {
-			charPos += 2
-			continue
-		}
-
-		if current.Len()+len(para) > maxChunkSize && current.Len() > 0 {
-			flush()
-		}
-
-		if current.Len() > 0 {
-			current.WriteString("\n\n")
-		}
-		current.WriteString(para)
-		charPos += len(para) + 2
-		paraIdx++
-	}
-	flush()
-
-	return chunks
-}
